requirements.txt
alembic==1.7.7
click==8.1.3
Faker==13.11.1
Flask==2.1.2
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
PyJWT==2.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
yapf==0.32.0
zipp==3.8.0
